requirements.txt
google-generativeai
python-dotenv
firebase-admin
streamlit